Refactor egui_web: break up into smaller files
